+++ cmplibrary/src/main/java/com/sourcepoint/cmplibrary/ConsentLib.java
@@ -7,10 +7,11 @@ import android.preference.PreferenceManager;
 import android.util.Log;
 import android.view.ViewGroup;
 
+import java.util.HashSet;
+
+import com.google.common.annotations.VisibleForTesting;
 import com.iab.gdpr_android.consent.VendorConsent;
 import com.iab.gdpr_android.consent.VendorConsentDecoder;
-
-import java.util.HashSet;
 
 /**
  * Entry point class encapsulating the Consents a giving user has given to one or several vendors.
@@ -79,12 +80,12 @@ public class ConsentLib {
 
     private static final String TAG = "ConsentLib";
     private static final String SP_PREFIX = "_sp_";
-    private static final String SP_PROPERTY_ID = SP_PREFIX + "site_id";
+    private static final String SP_SITE_ID = SP_PREFIX + "site_id";
     private final static String CUSTOM_CONSENTS_KEY = SP_PREFIX + "_custom_consents";
 
     private Activity activity;
-    private final String property;
-    private final int accountId, propertyId;
+    private final String siteName;
+    private final int accountId, siteId;
     private final ViewGroup viewGroup;
     private final Callback onMessageChoiceSelect, onConsentReady, onErrorOccurred;
     private Callback onMessageReady;
@@ -120,15 +121,15 @@ public class ConsentLib {
     /**
      * @return a new instance of ConsentLib.Builder
      */
-    public static ConsentLibBuilder newBuilder(Integer accountId, String property, Integer propertyId,String pmId ,Activity activity) {
-        return new ConsentLibBuilder(accountId, property, propertyId, pmId, activity);
+    public static ConsentLibBuilder newBuilder(Integer accountId, String siteName, Integer siteId,String pmId ,Activity activity) {
+        return new ConsentLibBuilder(accountId, siteName, siteId, pmId, activity);
     }
 
     ConsentLib(ConsentLibBuilder b) throws ConsentLibException.BuildException {
         activity = b.activity;
-        property = b.property;
+        siteName = b.siteName;
         accountId = b.accountId;
-        propertyId = b.propertyId;
+        siteId = b.siteId;
         encodedPMId = new EncodedParam("_sp_PMId",b.pmId);
         isShowPM = b.isShowPM;
         encodedAuthId = b.authId;
@@ -143,7 +144,7 @@ public class ConsentLib {
         // configurable time out
         defaultMessageTimeOut = b.defaultMessageTimeOut;
 
-        sourcePoint = new SourcePointClientBuilder(b.accountId, b.property + "/" + b.page, propertyId, b.staging)
+        sourcePoint = new SourcePointClientBuilder(b.accountId, b.siteName + "/" + b.page, siteId, b.staging)
                 .setStagingCampaign(b.stagingCampaign)
                 .setShowPM(b.isShowPM)
                 .setCmpDomain(b.cmpDomain)
@@ -456,12 +457,12 @@ public class ConsentLib {
     }
 
     private void loadAndStoreCustomVendorAndPurposeConsents(final String[] vendorIds, final OnLoadComplete callback) {
-        final String propertyIdKey = SP_PROPERTY_ID + "_" + accountId + "_" + property;
-        String propertyId = Integer.toString(this.propertyId);
+        final String siteIdKey = SP_SITE_ID + "_" + accountId + "_" + siteName;
+        String siteID = Integer.toString(siteId);
         SharedPreferences.Editor editor = sharedPref.edit();
-        editor.putString(propertyIdKey, propertyId);
+        editor.putString(siteIdKey, siteID);
         editor.apply();
-        sourcePoint.getCustomConsents(consentUUID, euconsent, propertyId, vendorIds, new OnLoadComplete() {
+        sourcePoint.getCustomConsents(consentUUID, euconsent, siteID, vendorIds, new OnLoadComplete() {
             @Override
             @SuppressWarnings("unchecked")
             public void onSuccess(Object result) {